% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{denoise}
\alias{denoise}
\title{Denoise SALT Observatory spectrum via trend filtering}
\usage{
denoise(sci, var, bpm, min_mask_width = 20, variability_bands = FALSE, ...)
}
\arguments{
\item{sci}{spectroscopic measurements.}

\item{var}{measurement variances.}

\item{bpm}{pixel masks.}

\item{min_mask_width}{(numeric) Parameter that determines the segmentation of
the spectrum into smaller pieces divided by gaps (many consecutive masked
spectral pixels). This is the minimum number of consecutive pixels that have
to be masked to cause a break in the spectrum.}

\item{variability_bands}{(Logical) If \code{TRUE} then variability bands are
computed via a bootstrap resampling algorithm based on the uncertainties
in the spectral pixel measurements.}

\item{B}{(integer) The number of bootstrap samples that are drawn in order to
estimate the trend filtering variability bands. Larger is better, but is more
computationally expensive. My trend filtering bootstrap implementation is set
up for parallel computing for speedups when multiple cores are available.
\code{B = 100} typically suffices for a most scenarios, but it's good to crank it
up a bit for a plot that's going to be published, so you get nice smooth
bands.}

\item{max_iter}{Maximum iterations allowed for the ADMM trend filtering
convex optimization (Ramdas and Tibshirani 2016). Increase this if the trend
filtering estimate does not appear to have fully converged to a reasonable
estimate of the spectral signal. An estimate that has not
fully converged will typically increasingly diverge above
or below the observed fluxes as you move from left to right.}

\item{obj_tol}{(integer) The tolerance used in the convex optimization
stopping criterion; when the relative change in the
objective function is less than this value, the algorithm
terminates. Decrease this if the trend
filtering estimate does not appear to have fully converged
to a reasonable estimate of the signal. An estimate that
has not fully converged will typically increasingly
diverge above or below the observed fluxes as you move
from left to right}
}
\value{
An object of class \code{\link[=denoise]{SALT_denoised}}. This is a list
with the following elements:
\item{x_eval}{Input grid used to evaluate the optimized trend filtering
estimate on.}
\item{tf_estimate}{Optimized trend filtering estimate, evaluated at
\code{x_eval}.}
}
\description{
Denoise SALT Observatory spectrum via trend filtering
}
\examples{
data(SALT_spectrum)
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
