% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{mask_intervals}
\alias{mask_intervals}
\title{Functions for handling spectral pixel masks}
\usage{
mask_intervals(df, min_mask_width = 20)
}
\arguments{
\item{df}{Input data frame}

\item{min_mask_width}{(numeric) Parameter that determines the segmentation
of the spectrum into smaller pieces divided by gaps (many consecutive masked
spectral pixels). This is the minimum number of consecutive pixels that have
to be masked to cause a break in the spectrum.}
}
\description{
Functions for handling spectral pixel masks
}
