% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{variability_bands.polarized_spectrum}
\alias{variability_bands.polarized_spectrum}
\title{Compute bootstrap variability bands for one of the denoised spectra}
\usage{
variability_bands.polarized_spectrum(obj, param, level = 0.95)
}
\arguments{
\item{obj}{An object of class \code{"polarized_spectrum"} produced by
\code{\link[=denoise_polarized_spectrum]{denoise_polarized_spectrum()}}.}

\item{param}{The denoised spectrum to compute variability bands for. One of
\code{c("I","Q","U","Q_norm","U_norm")}.}

\item{level}{The level of the pointwise variability bands. Defaults to
\code{level = 0.95}.}
}
\description{
Compute bootstrap variability bands for one of the denoised spectra
}
