% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{variability_bands}
\alias{variability_bands}
\title{Compute bootstrap variability bands for a denoised spectrum}
\usage{
variability_bands(obj, param, level = 0.95)
}
\arguments{
\item{obj}{An object of class \code{"polarized_spectrum"} produced by
\code{\link[=denoise_spectrum]{denoise_spectrum()}}.}

\item{param}{The denoised spectrum to compute variability bands for. One of
\code{c("I","Q","U","Q_norm","U_norm")}.}

\item{level}{The level of the pointwise variability bands. Defaults to
\code{level = 0.95}.}
}
\value{
A list of tibbles, with length equal to \code{obj$n_segments} and each
tibble with column set
\code{c("wavelength","bootstrap_lower_band","bootstrap_upper_band")}.
}
\description{
Compute bootstrap variability bands for a denoised spectrum
}
\examples{
data(polarized_spectrum_WR_star)

suppressMessages(library(dplyr))

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

flux <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) \%>\% select(1:3)
masks <- as_tibble(bpm$imDat)

\dontrun{
spec_denoised <- denoise_spectrum(
  wavelength,
  flux,
  variances,
  masks,
  compute_uncertainties = TRUE
)

bands <- variability_bands(spec_denoised, param = "Q_norm", level = 0.95)
}
}
