% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{break_spectrum}
\alias{break_spectrum}
\title{Break a spectrum into segments when chunks of many consecutively-masked
pixels are present}
\usage{
break_spectrum(df_full, break_at = 10, min_pix_segment = 10)
}
\arguments{
\item{df_full}{A tibble with minimal column set:
\code{c("wavelength","I","Q","U","I_mask","Q_mask","U_mask")}, and probably also
measurement variances for each of the Stokes parameters.}

\item{break_at}{A free parameter that controls the segmentation of a
spectrum. More precisely, \code{break_at} is the minimum number of
consecutively-masked spectral pixels that will trigger a break in the
spectrum. Defaults to \code{break_at = 10}.}

\item{min_pix_segment}{After the segmentation procedure is complete, it is
advisable to examine the resulting segments to ensure that each is
sufficiently long for its own denoising analysis. In particular, we discard
any segments that have less than \code{min_pix_segment} unmasked spectral pixels.
Defaults to \code{min_pix_segment = 10}.}
}
\value{
A list of tibbles corresponding to the segmented spectrum, as defined
by the above procedure and input argument choices. Each tibble in the list
has the same format as the input tibble \code{df_full}, less the mask columns.
The output tibbles only retain unmasked pixels, hence the removal of the
mask columns.
}
\description{
Break a spectrum into segments when chunks of many consecutively-masked
pixels are present
}
