% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{denoise_polarized_spectrum}
\alias{denoise_polarized_spectrum}
\title{Denoise a polarized spectrum and estimate the uncertainties}
\usage{
denoise_polarized_spectrum(
  wavelength,
  stokes,
  variances,
  masks,
  break_at = 10,
  min_pix_segment = 10,
  compute_uncertainties = FALSE,
  mc_cores = parallel::detectCores(),
  ...
)
}
\arguments{
\item{wavelength}{Vector of wavelength measurements.}

\item{stokes}{Polarized spectrum measurements, passed as a 3-column tibble,
data frame, or matrix, with the columns corresponding to the Stokes
parameters I, Q, and U, respectively.}

\item{variances}{Measurement variances, in a tibble, data frame, or matrix
with dimensions matching those of \code{stokes}.}

\item{masks}{Pixel masks, in a tibble, data frame, or matrix
with dimensions matching those of \code{stokes}. Nonzero elements flag bad pixels.}

\item{break_at}{A free parameter that controls the segmentation of a
spectrum. More precisely, \code{break_at} is the minimum number of
consecutively-masked spectral pixels that will trigger a break in the
spectrum. Defaults to \code{break_at = 10}.}

\item{min_pix_segment}{After the segmentation procedure is complete, the
resulting segments are examined to ensure that each is sufficiently long for
an independent denoising analysis. In particular, any segment that has less
than \code{min_pix_segment} unmasked spectral pixels is discarded. Defaults to
\code{min_pix_segment = 10}.}

\item{compute_uncertainties}{(Boolean) If \code{TRUE}, then variability bands are
computed for each of the denoised normalized Stokes spectra, via a parametric
bootstrap algorithm.}

\item{...}{Additional named arguments to be passed to
\link[trendfiltering:sure_trendfilter]{sure_trendfilter()} or
\link[trendfiltering:bootstrap_trendfilter]{bootstrap_trendfilter()}. The
argument \code{bootstrap_algorithm = "parametric"} is fixed due to the
design of SALT spectra, and we don't allow this argument to be manually
overridden. See the
\link[trendfiltering:bootstrap_trendfilter]{bootstrap_trendfilter()}
documentation and
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Politsch et al. (2020a)}
for details on why this particular bootstrap algorithm is appropriate for
SALT spectra.}
}
\value{
An object of class
\code{\link[=denoise_polarized_spectrum]{'polarized_spectrum_denoised'}}.
This is a list with the following elements:
}
\description{
Denoise a polarized spectrum and estimate the uncertainties
}
\examples{
data(polarized_spectrum_WR_star)

suppressMessages(library(dplyr))

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

stokes <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) \%>\% select(1:3)
masks <- as_tibble(bpm$imDat)

spec_denoised <- denoise_polarized_spectrum(wavelength, stokes, variances, masks)
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
