% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{denoise_polarized_spectrum}
\alias{denoise_polarized_spectrum}
\title{Denoise a SALT Observatory spectrum via quadratic trend filtering and compute
uncertainties via a bootstrap}
\usage{
denoise_polarized_spectrum(
  wavelength,
  stokes,
  variances,
  masks,
  spectrum_break = 10,
  compute_uncertainties = FALSE,
  ...
)
}
\arguments{
\item{stokes}{Spectropolarimetric measurements in a Stokes parametrization,
passed as a 3-column matrix, with the columns corresponding to the I, Q, U
Stokes parameters, respectively.}

\item{variances}{Measurement variances, in a matrix with dimensions matching
those of \code{stokes}.}

\item{masks}{Pixel masks, in a matrix with dimensions matching those
of \code{stokes}. Nonzero elements flag bad pixels.}

\item{spectrum_break}{Parameter controlling the segmentation of a spectrum
into ``sub-spectra'', which are then denoised independently. More precisely,
\code{spectrum_break} is the minimum number of consecutively-masked spectral
pixels required to trigger a break in the spectrum. Defaults to
`spectrum_break = 10`.}

\item{compute_uncertainties}{(Boolean) If \code{TRUE} then variability bands are
computed for each of the denoised normalized Stokes spectra, via a parametric
bootstrap algorithm.}

\item{...}{Additional named arguments to be passed to
\code{\link[trendfiltering:sure_trendfilter]{trendfiltering::sure_trendfilter()}} or
\code{\link[trendfiltering:bootstrap_trendfilter]{trendfiltering::bootstrap_trendfilter()}}.}
}
\value{
An object of class
\code{\link[=denoise_polarized_spectrum]{'polarized_spectrum_denoised'}}.
This is a list with the following elements:
}
\description{
Denoise a SALT Observatory spectrum via quadratic trend filtering and compute
uncertainties via a bootstrap
}
\examples{
data(SALT_spectrum)

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

stokes <- as_tibble(sci$imDat) \%>\%
  rename_with(function(.cols) c("I", "Q", "U"))
weights <- as_tibble(var$imDat) \%>\%
  select(1:3) \%>\%
  rename_with(function(.cols) c("I_weights", "Q_weights", "U_weights"))
masks <- as_tibble(bpm$imDat) \%>\%
  rename_with(function(.cols) c("I_mask", "Q_mask", "U_mask"))
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
