% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.R
\name{denoise_spectrum}
\alias{denoise_spectrum}
\title{Denoise a SALT Observatory spectrum via quadratic trend filtering and compute
uncertainties via a Gaussian bootstrap}
\usage{
denoise_spectrum(
  sci,
  var,
  bpm,
  min_mask_width = 20,
  compute_uncertainties = FALSE,
  ...
)
}
\arguments{
\item{sci}{spectroscopic measurements.}

\item{var}{measurement variances.}

\item{bpm}{pixel masks.}

\item{min_mask_width}{Parameter that controls the segmentation of a spectrum
into ``sub-spectra'', which are then denoised independently. More precisely,
\code{min_mask_width} is the minimum number of consecutively-masked spectral
pixels in order to trigger a break in the spectrum. Defaults to
`min_mask_width = 20`.}

\item{compute_uncertainties}{(Boolean) If \code{TRUE} then variability bands are
computed via a parametric bootstrap algorithm based on the uncertainties
in the spectral pixel measurements.}

\item{nx_eval}{Integer. If nothing is passed to \code{x_eval}, then it is defined
as \code{x_eval = seq(min(x), max(x), length = nx_eval)}.}

\item{x_eval}{(Optional) A grid of inputs to evaluate the optimized trend
filtering estimate on. May be ignored, in which case the grid is determined
by \code{nx_eval}.}

\item{B}{(Integer) The number of bootstrap samples that are drawn in order to
estimate the trend filtering variability bands. Larger is better, but is more
computationally expensive. My trend filtering bootstrap implementation is set
up for parallel computing for speedups when multiple cores are available.
\code{B = 100} typically suffices for a most scenarios, but it's good to crank it
up a bit for a plot that's going to be published, so you get nice smooth
bands.}
}
\value{
An object of class \code{\link[=denoise_spectrum]{'SALT_spec'}}. This is a
list with the following elements:
\item{x_eval}{Input grid used to evaluate the optimized trend filtering
estimate on.}
\item{tf_estimate}{Optimized trend filtering estimate, evaluated at
\code{x_eval}.}
\item{validation_method}{"SURE"}
\item{lambdas}{Vector of hyperparameter values evaluated in the grid search
(always returned in descending order).}
\item{edfs}{Vector of effective degrees of freedom for all trend filtering
estimators fit during validation.}
\item{generalization_errors}{Vector of SURE generalization error estimates,
corresponding to the descending-ordered \code{lambdas} vector.}
\item{lambda_min}{Hyperparameter value that minimizes the SURE generalization
error curve.}
\item{i_min}{Index of \code{lambdas} that minimizes the SURE error curve.}
\item{edf_min}{Effective degrees of freedom of the optimized trend
filtering estimator.}
\item{n_iter}{The number of iterations needed for the ADMM algorithm to
converge within the given tolerance, for each hyperparameter value. If many
of these are exactly equal to \code{max_iter}, then their solutions have not
converged with the tolerance specified by \code{obj_tol}. In which case, it is
often prudent to increase \code{max_iter}.}
\item{x}{Vector of observed inputs.}
\item{y}{Vector of observed outputs.}
\item{weights}{Weights for the observed outputs, defined as the reciprocal
variance of the additive noise that contaminates the signal.}
\item{fitted_values}{Optimized trend filtering estimate, evaluated at the
observed inputs \code{x}.}
\item{residuals}{\code{residuals = y - fitted_values}}
}
\description{
Denoise a SALT Observatory spectrum via quadratic trend filtering and compute
uncertainties via a Gaussian bootstrap
}
\examples{
data(SALT_spectrum)
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
