% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_bands.R
\name{variability_bands}
\alias{variability_bands}
\title{Quantify the statistical uncertainty in a denoised spectrum by computing
bootstrap variability bands}
\usage{
variability_bands(obj, param, level = 0.95)
}
\arguments{
\item{obj}{An object of class \code{"polarized_spectrum"}, produced by
\code{\link[=denoise_spectrum]{denoise_spectrum()}}.}

\item{param}{A string specifying which spectrum to compute variability bands
for. One of \code{c("I","Q","U","Q_norm","U_norm")}.}

\item{level}{The level of the pointwise variability bands. Defaults to
\code{level = 0.95}.}
}
\value{
A list of \code{obj$n_segments} tibbles, each with the column set
\code{c("wavelength","bootstrap_lower_band","bootstrap_upper_band")}.
}
\description{
See the parametric bootstrap algorithm in \href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Politsch et al. (2020a)} for details.
}
\examples{
data(polarized_spectrum_WR_star)
library(dplyr)

wavelength <- seq(
  from = sci$axDat$crval[1],
  by = sci$axDat$cdelt[1],
  length = sci$axDat$len[1]
)

flux <- as_tibble(sci$imDat)
variances <- as_tibble(var$imDat) \%>\% select(1:3)
masks <- as_tibble(bpm$imDat)

spec_denoised <- denoise_spectrum(
  wavelength,
  flux,
  variances,
  masks,
  compute_uncertainties = TRUE
)

bands <- variability_bands(spec_denoised, param = "Q_norm", level = 0.95)
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
\seealso{
\code{\link[=denoise_spectrum]{denoise_spectrum()}}
}
